+++ models/commerce/tax_rates.py
@@ -1,4 +1,4 @@
-from sqlalchemy import String, Float, Boolean, DateTime, func, Index
+from sqlalchemy import String, Numeric, Boolean, DateTime, func, Index
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 from datetime import datetime, datetime as dt
@@ -25,7 +25,7 @@ class TaxRate(Base):
     country_name: Mapped[str] = mapped_column(String(100))
     province_code: Mapped[Optional[str]] = mapped_column(String(10), nullable=True)  # State/Province code
     province_name: Mapped[Optional[str]] = mapped_column(String(100), nullable=True)
-    tax_rate: Mapped[float] = mapped_column(Float)  # Tax rate as decimal (e.g., 0.13 for 13%)
+    tax_rate: Mapped[float] = mapped_column(Numeric(5, 4))  # Tax rate as decimal (e.g., 0.13 for 13%)
     tax_name: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)  # e.g., "GST", "VAT", "Sales Tax"
     is_active: Mapped[bool] = mapped_column(Boolean, default=True)
     effective_date: Mapped[Optional[datetime]] = mapped_column(DateTime, nullable=True)
